Drop all former executables […]
Maintaining past artworks takes a lot of effort. Instead, they should be
regarded as works that are based on a certain point in time of this library.
Should the need arise to rebase them onto master that is most likely much easier
(and allows some refactoring even) than constantly maintaining not only their
own code, but also their dependencies when updating LTS releases. Speaking of
dependencies: it’s much easier to add dedicated dependencies to a branch (low
barrier of entry) than the whole repo (where it should have a *lot* of utility
to justify its existence).