Remove reverse function […]
During refactoring of the old code, I checked whether the MD5 hash of the
doctest-generated SVG file changed. It turned out it *did* at some point, namely
because the old algorithm generated the output in reverse order, because it put
newly created points ontop of the result (list) stack. To maintain 100% MD5
consistency, I put in an artificial reverse in the now fixed-and-lazy solution.
Now that I’m done I can finally take it out :-)